#329 Use file reader

+++ neclib/utils/file_utils.py
@@ -8,18 +8,24 @@ from urllib.request import urlopen
 
 
 @overload
-def read_file(name: str, /, saveto: None = None) -> Union[str, bytes]:
+def read_file(
+    name: str, /, saveto: None, overwrite: bool, localonly: bool
+) -> Union[str, bytes]:
     ...
 
 
 @overload
-def read_file(name: str, /, saveto: os.PathLike) -> None:
+def read_file(
+    name: str, /, saveto: os.PathLike, overwrite: bool, localonly: bool
+) -> None:
     ...
 
 
-def read_file(name, /, saveto=None):
+def read_file(name, /, saveto=None, overwrite=False, localonly=False):
     name = str(name)
     if urlparse(name).scheme:
+        if localonly and (urlparse(name).scheme != "file"):
+            raise FileNotFoundError(f"{name} appears to be a remote file.")
         with urlopen(name) as response:
             contents = response.read().decode("utf-8")
     elif Path(name).exists():
@@ -29,10 +35,15 @@ def read_file(name, /, saveto=None):
         except UnicodeDecodeError:
             contents = path.read_bytes()
     else:
-        raise ValueError(f"Cannot find {name!r}.")
+        raise FileNotFoundError(f"Cannot find {name!r}.")
 
     if saveto is not None:
         save_path = Path(saveto)
+        if save_path.exists() and not overwrite:
+            raise FileExistsError(f"{save_path} already exists.")
+
+        save_path.parent.mkdir(parents=True, exist_ok=True)
+        save_path.touch(exist_ok=True)
         try:
             save_path.write_text(contents)
         except TypeError:
